refactor(api): move user-to-create repository
