+++ api/scripts/create-users-accounts-for-contest.js
@@ -3,10 +3,10 @@ import * as url from 'node:url';
 import bluebird from 'bluebird';
 
 import { disconnect } from '../db/knex-database-connection.js';
+import { userToCreateRepository } from '../src/identity-access-management/infrastructure/repositories/user-to-create.repository.js';
 import { cryptoService } from '../src/shared/domain/services/crypto-service.js';
 import * as userService from '../src/shared/domain/services/user-service.js';
 import * as authenticationMethodRepository from '../src/shared/infrastructure/repositories/authentication-method-repository.js';
-import * as userToCreateRepository from '../src/shared/infrastructure/repositories/user-to-create-repository.js';
 import { parseCsvWithHeader } from './helpers/csvHelpers.js';
 
 function prepareDataForInsert(rawUsers) {
