+++ api/tests/integration/domain/usecases/create-and-reconcile-user-to-organization-learner_test.js
@@ -12,6 +12,7 @@ import * as userReconciliationService from '../../../../lib/domain/services/user
 import { createAndReconcileUserToOrganizationLearner } from '../../../../lib/domain/usecases/create-and-reconcile-user-to-organization-learner.js';
 import * as campaignRepository from '../../../../lib/infrastructure/repositories/campaign-repository.js';
 import * as organizationLearnerRepository from '../../../../lib/infrastructure/repositories/organization-learner-repository.js';
+import { userToCreateRepository } from '../../../../src/identity-access-management/infrastructure/repositories/user-to-create.repository.js';
 import { EntityValidationError } from '../../../../src/shared/domain/errors.js';
 import { cryptoService } from '../../../../src/shared/domain/services/crypto-service.js';
 import * as userService from '../../../../src/shared/domain/services/user-service.js';
@@ -19,7 +20,6 @@ import * as passwordValidator from '../../../../src/shared/domain/validators/pas
 import * as userValidator from '../../../../src/shared/domain/validators/user-validator.js';
 import * as authenticationMethodRepository from '../../../../src/shared/infrastructure/repositories/authentication-method-repository.js';
 import * as userRepository from '../../../../src/shared/infrastructure/repositories/user-repository.js';
-import * as userToCreateRepository from '../../../../src/shared/infrastructure/repositories/user-to-create-repository.js';
 import { catchErr, databaseBuilder, expect } from '../../../test-helper.js';
 
 describe('Integration | UseCases | create-and-reconcile-user-to-organization-learner', function () {
