+++ api/tests/shared/integration/domain/services/user-service_test.js
@@ -4,10 +4,10 @@ const { pick } = lodash;
 import { OrganizationLearnerNotFound } from '../../../../../lib/domain/errors.js';
 import * as organizationLearnerRepository from '../../../../../lib/infrastructure/repositories/organization-learner-repository.js';
 import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../../../src/identity-access-management/domain/constants/identity-providers.js';
+import { userToCreateRepository } from '../../../../../src/identity-access-management/infrastructure/repositories/user-to-create.repository.js';
 import * as userService from '../../../../../src/shared/domain/services/user-service.js';
 import * as authenticationMethodRepository from '../../../../../src/shared/infrastructure/repositories/authentication-method-repository.js';
 import * as userRepository from '../../../../../src/shared/infrastructure/repositories/user-repository.js';
-import * as userToCreateRepository from '../../../../../src/shared/infrastructure/repositories/user-to-create-repository.js';
 import { catchErr, databaseBuilder, domainBuilder, expect, knex } from '../../../../test-helper.js';
 
 describe('Integration | Domain | Services | user-service', function () {
